+++ src/features/product/productPost/tabs/components/ListItem.tsx
@@ -1,4 +1,16 @@
-export default function ListItem({ tabName, activeTab, setActiveTab, singleProduct }) {
+import type { Dispatch, SetStateAction } from 'react';
+
+import type { Product } from '@typings/products/product';
+import type { ActiveTab } from '@typings/products/tab';
+
+interface ListItemProps {
+  tabName: ActiveTab;
+  activeTab: ActiveTab;
+  setActiveTab: Dispatch<SetStateAction<ActiveTab>>;
+  product?: Product;
+}
+
+export default function ListItem({ tabName, activeTab, setActiveTab, product }: ListItemProps) {
   const activeBgColorClass = activeTab === tabName ? 'bg-pearl-bush' : 'bg-pampas hover:bg-pearl-bush';
 
   return (
@@ -12,7 +24,7 @@ export default function ListItem({ tabName, activeTab, setActiveTab, singleProdu
       >
         <span className="flex gap-x-1 text-xs tracking-xwide text-shark uppercase">
           {tabName}
-          {tabName === 'reviews' && <span>({singleProduct.reviewCount || 0})</span>}
+          {tabName === 'reviews' && <span>({product?.reviewCount || 0})</span>}
         </span>
       </button>
     </li>